requirements.txt
Django==4.1.7
django-filter==22.1
psycopg2-binary==2.9.5
requests==2.28.1
